refactor(errors): rename error_handling module to errors
