+++ component/template/src/errors.rs
